+++ R/provider-google.R
@@ -840,6 +840,46 @@ gemini_client <- function() {
   )
 }
 
+# Token counting ----------------------------------------------------------
+
+# https://ai.google.dev/api/tokens
+method(count_tokens, ProviderGoogleGemini) <- function(
+  provider,
+  turns = list(),
+  tools = list(),
+  type = NULL
+) {
+  req <- base_request(provider)
+  req <- req_url_path_append(
+    req,
+    "models",
+    paste0(provider@model, ":", "countTokens")
+  )
+
+  body <- chat_body(
+    provider,
+    stream = FALSE,
+    turns = turns,
+    tools = tools,
+    type = type
+  )
+
+  keep <- c("contents", "tools", "systemInstruction", "generationConfig")
+  generate_content_request <- body[intersect(names(body), keep)]
+  generate_content_request$model <- paste0("models/", provider@model)
+
+  req <- req_body_json(
+    req,
+    list(
+      generateContentRequest = generate_content_request
+    )
+  )
+  req <- req_headers(req, !!!provider@extra_headers)
+
+  resp <- req_perform(req)
+  resp_body_json(resp)$totalTokens %||% 0L
+}
+
 # Pricing ----------------------------------------------------------------------
 
 # Models -----------------------------------------------------------------------